refact: Estrutura dos testes
